update option's page layout
