+++ components/ui/button.tsx
@@ -1,3 +1,4 @@
+// biome-ignore lint/nursery/noRestrictedImports: <explanation>
 import { Slot } from "@radix-ui/react-slot"
 import { type VariantProps, cva } from "class-variance-authority"
 import * as React from "react"
