+++ components/ui/switch.tsx
@@ -1,3 +1,4 @@
+// biome-ignore lint/nursery/noRestrictedImports: <explanation>
 import * as SwitchPrimitives from "@radix-ui/react-switch"
 import * as React from "react"
